+++ src/main/java/insane96mcp/iguanatweaksreborn/module/farming/plantsgrowth/PlantGrowthModifiers.java
@@ -1,7 +1,10 @@
 package insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth;
 
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
+import insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth.modifier.CorrectBiomeModifier;
+import insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth.modifier.NightTimeModifier;
 import insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth.modifier.SunlightModifier;
+import insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth.modifier.WrongBiomeModifier;
 import net.minecraft.resources.ResourceLocation;
 
 import java.lang.reflect.Type;
@@ -10,7 +13,13 @@ import java.util.Map;
 
 public class PlantGrowthModifiers {
     public static final Map<ResourceLocation, Type> MODIFIERS = new HashMap<>();
-    public static final Type SUNLIGHT = registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "sunlight"), SunlightModifier.class);
+
+    public static void init() {
+        registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "sunlight"), SunlightModifier.class);
+        registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "night_time"), NightTimeModifier.class);
+        registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "correct_biome"), CorrectBiomeModifier.class);
+        registerModifier(new ResourceLocation(IguanaTweaksReborn.MOD_ID, "wrong_biome"), WrongBiomeModifier.class);
+    }
 
     public static Type registerModifier(ResourceLocation id, Type modifier) {
         MODIFIERS.put(id, modifier);