+++ src/main/java/insane96mcp/iguanatweaksreborn/module/farming/plantsgrowth/PlantGrowthMultiplier.java
@@ -19,13 +19,6 @@ public class PlantGrowthMultiplier {
 	public final IdTagMatcher block;
 	private final float growthMultiplier;
 	protected final List<PlantGrowthModifier> modifiers = new ArrayList<>();
-	/*private float noSunlightMultiplier = 1f;
-	private int minSunlightRequired = 0;
-	private float nightTimeMultiplier = 1f;
-	private List<IdTagMatcher> correctBiomes = new ArrayList<>();
-	private boolean invertCorrectBiomes = false;
-	private float wrongBiomeMultiplier = 1f;
-	public final List<SeasonMultiplier> seasonsMultipliers = new ArrayList<>();*/
 
 	private PlantGrowthMultiplier(IdTagMatcher block, float growthMultiplier, List<PlantGrowthModifier> modifiers) {
 		this.block = block;
@@ -33,39 +26,6 @@ public class PlantGrowthMultiplier {
 		this.modifiers.addAll(modifiers);
 	}
 
-	/*@JsonAdapter(SeasonMultiplier.Serializer.class)
-	public record SeasonMultiplier(Object season, float multiplier) {
-		public static final java.lang.reflect.Type LIST_TYPE = new TypeToken<ArrayList<SeasonMultiplier>>(){}.getType();
-
-		public static class Serializer implements JsonDeserializer<SeasonMultiplier>, JsonSerializer<SeasonMultiplier> {
-			@Override
-			public SeasonMultiplier deserialize(JsonElement json, java.lang.reflect.Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-				JsonObject jObject = json.getAsJsonObject();
-				Object season = null;
-				if (jObject.has("season")) {
-					if (!ModList.get().isLoaded("sereneseasons"))
-						throw new JsonParseException("Tried deserializing Plant Growth Modifier `season_multipliers` but Serene Season is not installed");
-					season = SereneSeasons.deserializeSeason(jObject, "season");
-				}
-				float multiplier = GsonHelper.getAsFloat(jObject, "multiplier");
-
-				return new SeasonMultiplier(season, multiplier);
-			}
-
-			@Override
-			public JsonElement serialize(SeasonMultiplier src, java.lang.reflect.Type typeOfSrc, JsonSerializationContext context) {
-				JsonObject jObject = new JsonObject();
-				if (src.season != null) {
-					if (!ModList.get().isLoaded("sereneseasons"))
-						throw new JsonParseException("Tried serializing Plant Growth Modifier `season_multipliers` but Serene Season is not installed");
-					jObject.addProperty("season", SereneSeasons.serializeSeason(src.season));
-				}
-				jObject.addProperty("multiplier", src.multiplier);
-				return jObject;
-			}
-		}
-	}
-*/
 	/**
 	 * Returns >=1 for the chance 1 in this to grow
 	 * Values between 0 and 1 have no effect
@@ -79,34 +39,6 @@ public class PlantGrowthMultiplier {
 			multiplier *= modifier.getMultiplier(state, level, pos);
 		}
 		return multiplier;
-		/*int dayTime = (int) (level.dayTime() % 24000);
-		if (dayTime >= 12786 && dayTime < 23216)
-			multiplier *= this.nightTimeMultiplier;
-		Holder<Biome> biome = level.getBiome(pos);
-		if (!this.correctBiomes.isEmpty()) {
-			boolean isInCorrectBiome = false;
-			for (IdTagMatcher correctBiome : this.correctBiomes) {
-				if (correctBiome.matchesBiome(biome)) {
-					isInCorrectBiome = true;
-					break;
-				}
-			}
-			//If is not in correct biome or if is but the correct biomes becomes wrong biomes
-			if (!this.invertCorrectBiomes && !isInCorrectBiome
-					|| this.invertCorrectBiomes && isInCorrectBiome) {
-				multiplier *= this.wrongBiomeMultiplier;
-			}
-		}
-		if (!this.seasonsMultipliers.isEmpty()) {
-			for (SeasonMultiplier seasonMultiplier : this.seasonsMultipliers) {
-				if (SereneSeasons.doesSeasonMatch(seasonMultiplier.season, level)) {
-					multiplier *= seasonMultiplier.multiplier;
-					break;
-				}
-			}
-		}
-
-		return multiplier;*/
 	}
 
 	public static final java.lang.reflect.Type LIST_TYPE = new TypeToken<ArrayList<PlantGrowthMultiplier>>(){}.getType();