+++ src/main/java/insane96mcp/iguanatweaksreborn/module/farming/plantsgrowth/modifier/WrongBiomeModifier.java
@@ -1,0 +1,45 @@
+package insane96mcp.iguanatweaksreborn.module.farming.plantsgrowth.modifier;
+
+import com.google.gson.*;
+import com.google.gson.annotations.JsonAdapter;
+import insane96mcp.insanelib.data.IdTagMatcher;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
+import net.minecraft.util.GsonHelper;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.block.state.BlockState;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@JsonAdapter(WrongBiomeModifier.Serializer.class)
+public class WrongBiomeModifier extends AbstractBiomeModifier {
+    protected WrongBiomeModifier(float multiplier, List<IdTagMatcher> biomes) {
+        super(multiplier, biomes);
+    }
+
+    @Override
+    public float getMultiplier(BlockState state, Level level, BlockPos pos) {
+        Holder<Biome> currentBiome = level.getBiome(pos);
+        for (IdTagMatcher biome : this.biomes) {
+            if (biome.matchesBiome(currentBiome))
+                return this.multiplier;
+        }
+        return 1f;
+    }
+
+    public static class Serializer implements JsonDeserializer<WrongBiomeModifier> {
+        @Override
+        public WrongBiomeModifier deserialize(JsonElement json, java.lang.reflect.Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+            JsonObject jObject = json.getAsJsonObject();
+            List<IdTagMatcher> biomes = new ArrayList<>();
+            JsonArray aCorrectBiomes = GsonHelper.getAsJsonArray(json.getAsJsonObject(), "biomes");
+            for (JsonElement jsonElement : aCorrectBiomes) {
+                IdTagMatcher biome = context.deserialize(jsonElement, IdTagMatcher.class);
+                biomes.add(biome);
+            }
+            return new WrongBiomeModifier(GsonHelper.getAsFloat(jObject, "multiplier"), biomes);
+        }
+    }
+}